Split native_hooks parser and hide internal types
Move TOML deserialization into `native_hooks::parser` (kept private).
Public API of the module shrinks to just `install(core, data,
scan_ranges)`: hashing the binary, picking the matching entry, scanning
patterns, patching SVCs, and registering the dispatcher all happen
inside. `Entry`, `Hook`, `HookKind`, `InlineCopy`, `PatternHook`,
`PatternToken`, `CaptureName`, `PatternHookKind`, the `md5` helper, the
`native_hooks()` parser entry point, and the SVC base/max constants are
no longer exposed.

Replace per-hook SVC categories with a single category (0x80) plus a
`BTreeMap<pc, HookKind>` registry the unified `handle_native_hook`
dispatcher consults. PC-based lookup removes the artificial
`NATIVE_HOOK_MAX` cap on installed hooks and collapses the five
per-kind handlers into one match.

KTF's call simplifies to `install(core, data, &[(IMAGE_BASE,
data.len() as u32)])` — the lookup logic is no longer duplicated at
the caller.
